requirements.txt
pyrogram
tgcrypto
ruamel.yaml
